+++ src/main/java/com/cta4j/bus/vehicle/model/Vehicle.java
@@ -9,7 +9,7 @@ import java.util.Objects;
  * Represents a vehicle.
  *
  * @param id the unique identifier of this vehicle
- * @param route the alphanumeric designator of the route that is currently being serviced by this vehicle
+ * @param routeId the alphanumeric designator of the route that is currently being serviced by this vehicle
  * @param destination the destination of the trip being serviced by this vehicle (e.g. "Austin")
  * @param coordinates the current coordinates of this vehicle
  * @param delayed whether this vehicle is currently delayed
@@ -19,7 +19,7 @@ import java.util.Objects;
 public record Vehicle(
     String id,
 
-    String route,
+    String routeId,
 
     String destination,
 
@@ -33,17 +33,17 @@ public record Vehicle(
      * Constructs a {@code Vehicle}.
      *
      * @param id the unique identifier of the vehicle
-     * @param route the alphanumeric designator of the route that is currently being serviced by the vehicle
+     * @param routeId the alphanumeric designator of the route that is currently being serviced by the vehicle
      * @param destination the destination of the trip being serviced by the vehicle (e.g. "Austin")
      * @param coordinates the current coordinates of the vehicle
      * @param delayed whether the vehicle is currently delayed
      * @param metadata the metadata associated with the vehicle
-     * @throws NullPointerException if {@code id}, {@code route}, {@code destination}, {@code coordinates}, or
+     * @throws NullPointerException if {@code id}, {@code routeId}, {@code destination}, {@code coordinates}, or
      * {@code metadata} is {@code null}
      */
     public Vehicle {
         Objects.requireNonNull(id);
-        Objects.requireNonNull(route);
+        Objects.requireNonNull(routeId);
         Objects.requireNonNull(destination);
         Objects.requireNonNull(coordinates);
         Objects.requireNonNull(metadata);